requirements.txt
Package                  Version
------------------------ ------------
albumentations           1.0.3
annotated-types          0.6.0
certifi                  2024.2.2
charset-normalizer       3.3.2
contourpy                1.2.1
cycler                   0.12.1
filelock                 3.9.0
fonttools                4.51.0
fsspec                   2024.3.1
huggingface-hub          0.22.2
idna                     3.7
imageio                  2.34.0
Jinja2                   3.1.2
joblib                   1.4.0
kiwisolver               1.4.5
lazy_loader              0.4
lightning-utilities      0.11.2
MarkupSafe               2.1.3
matplotlib               3.8.4
mpmath                   1.3.0
networkx                 3.3
numpy                    1.26.4
nvidia-cublas-cu11       11.11.3.6
nvidia-cuda-cupti-cu11   11.8.87
nvidia-cuda-nvrtc-cu11   11.8.89
nvidia-cuda-runtime-cu11 11.8.89
nvidia-cudnn-cu11        8.7.0.84
nvidia-cufft-cu11        10.9.0.58
nvidia-curand-cu11       10.3.0.86
nvidia-cusolver-cu11     11.4.1.48
nvidia-cusparse-cu11     11.7.5.86
nvidia-nccl-cu11         2.19.3
nvidia-nvtx-cu11         11.8.86
opencv-python            4.5.5.64
opencv-python-headless   4.9.0.80
packaging                24.0
pandas                   2.2.1
pillow                   10.3.0
pip                      24.0
pydantic                 2.7.0
pydantic_core            2.18.1
pyparsing                3.1.2
python-dateutil          2.9.0.post0
pytz                     2024.1
PyYAML                   6.0.1
requests                 2.31.0
safetensors              0.4.3
scikit-image             0.23.1
scikit-learn             1.4.2
scipy                    1.13.0
setuptools               69.2.0
six                      1.16.0
sympy                    1.12
threadpoolctl            3.4.0
tifffile                 2024.4.18
timm                     0.9.16
torch                    2.2.2+cu118
torchaudio               2.2.2+cu118
torchmetrics             1.3.2
torchvision              0.17.2+cu118
tqdm                     4.66.2
triton                   2.2.0
typing_extensions        4.11.0
tzdata                   2024.1
urllib3                  2.2.1
wheel                    0.43.0